+++ package.json
@@ -38,7 +38,6 @@
     "ts-loader": "^9.2.6",
     "ts-node": "^10.7.0",
     "typescript": "^4.5.5",
-    "vite": "^2.9.1",
-    "vite-plugin-html": "^3.2.0"
+    "vite": "^2.9.1"
   }
 }
