--- src/pages/Devtools/index.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8">
-</head>
-<body>
-<script src="devtools.js"></script>
-</body>
-</html>
