--- src/pages/Newtab/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="UTF-8" />
-    <title>New tab</title>
-</head>
-
-<body>
-<div id="app-container"></div>
-</body>
-</html>
