+++ src/pages/background/index.ts
