+++ src/pages/content/index.ts
@@ -1,1 +1,2 @@
+import "./style.css";
 console.log("content loaded");
