+++ src/pages/devtools/index.html
@@ -2,10 +2,9 @@
 <html lang="ko">
 <head>
     <meta charset="UTF-8" />
-    <title>Devtools Panel</title>
+    <title>Devtools</title>
 </head>
-
 <body>
-<div id="app-container"></div>
+<script type="module" src="./index.tsx"></script>
 </body>
 </html>
