+++ src/pages/newtab/Newtab.css
