+++ src/pages/newtab/Newtab.scss
