+++ src/pages/options/Options.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "@pages/Options/Options.css";
+import "@pages/options/Options.css";
 
 const Options: React.FC = () => {
   return <div className="OptionsContainer">Options</div>;
