+++ src/pages/options/index.html
@@ -2,10 +2,11 @@
 <html lang="ko">
 <head>
     <meta charset="UTF-8" />
-    <title>Popup</title>
+    <title>Options</title>
 </head>
 
 <body>
 <div id="app-container"></div>
+<script type="module" src="./index.tsx"></script>
 </body>
 </html>
