+++ src/pages/panel/Panel.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "@pages/Panel/Panel.css";
+import "@pages/panel/Panel.css";
 
 const Panel: React.FC = () => {
   return (
