+++ src/pages/popup/Popup.tsx
@@ -1,15 +1,14 @@
 import React from "react";
 import logo from "@assets/img/logo.svg";
-import "@pages/Newtab/Newtab.css";
-import "@pages/Newtab/Newtab.scss";
+import "@pages/popup/Popup.css";
 
-const Newtab = () => {
+const Popup = () => {
   return (
     <div className="App">
       <header className="App-header">
         <img src={logo} className="App-logo" alt="logo" />
         <p>
-          Edit <code>src/pages/Newtab/Newtab.tsx</code> and save to reload.
+          Edit <code>src/pages/Popup/Popup.jsx</code> and save to reload.
         </p>
         <a
           className="App-link"
@@ -19,10 +18,9 @@ const Newtab = () => {
         >
           Learn React!
         </a>
-        <h6>The color of this paragraph is defined using SASS.</h6>
       </header>
     </div>
   );
 };
 
-export default Newtab;
+export default Popup;
