+++ src/pages/popup/index.html
@@ -7,5 +7,6 @@
 
 <body>
 <div id="app-container"></div>
+<script type="module" src="./index.tsx"></script>
 </body>
 </html>
